Add files via upload
The goal of the university management code is to make managing student data easier. A'students' table is created by the application using a Mysql that stores important data such names, roll numbers, and departments. Functions to add student information to the database and retrieve it for viewing are included in the code. Through a simple Python program, this succinct script offers a basis for effectively handling and dealing with student data in an academic setting.